pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
